+++ spec/spec_helper.rb
@@ -152,4 +152,5 @@ RSpec.configure do |config|
   #     Bullet.end_request
   #   end
   # end
+  config.include ActionCable::TestHelper
 end